Refactor DXF types, enums, and viewport handling
Standardize use of fixed-width integer types for DXF group codes, flags, and color indices. Refactor enums to use enum class and explicit underlying types. Replace DXF color table with a spec-faithful, Doxygen-documented aciCanonicalRgbTable. Redesign EoDxfVPort with descriptive member names and geometry types; update all parsing and writing logic accordingly. Add GetInt16() to EoDxfReader and update entity/table parsing to use it. Improve documentation and code comments, and update copilot-instructions.md to emphasize verbose, self-documenting code. Use named constants and strong types for handles, colors, and flags throughout for clarity and maintainability.

+++ Core Libraries/EoDxfLib/EoDxfHeader.h
@@ -14,7 +14,7 @@ class EoDxfWriter;
  */
 class EoDxfHeader {
  public:
-  EoDxfHeader() { m_version = EoDxf::Version::AC1021; }
+  EoDxfHeader() { m_version = static_cast<int>(EoDxf::Version::AC1021); }
 
   EoDxfHeader(const EoDxfHeader& other);
